Making and lexing
I started work on getting things made so that I can test the program
just by typing the work "make" in the terminal, but I've encountered
some issues with the making of the test executable. It saying that a
function is undefined, although it looks to be defined. I will have to
do some searching to fix that.

I've implemented some more components of the lexer, not yet the numbers
but I do have a plan to implement it and left it at the end
of the lexer in comment form.