Started work on the parser
I need to implement a way to determine which branch the program
should take, and generate the AST on.
I will also need to implement the way for the expression function to
setup for the lexer.
I was thinking for the AST I could use a number to determine the branch
but so far that has not worked out, although I think there is a way to
do it this way.