Started message handler
There needs to be a message handler so that the compiler can give
valuable information back to the programmer when the program is running.
I will make it so that the programmer can easily set the function that
should be called to handle the message, and an enum of the message
levels and the message level names.
I need to implement the message calling and the default message handler
itself.