Update
I've added comments to the functions and variables that need comments.
I've done some testing and implemented the ability for numbers
which go along the lines of .45645 and for empty parentheses.
I've implemented the message handler, if debug is off and the message
is a debug message, it will return before doing
a call to the custom handler or the default handler.
The compiler does not stop when it encounters an error by default.
This can be implemented with the use of a custom message handler.
The compiler_params structure will either be removed and replaced by
just variables, or be revised.

I am also thinking about integrating the compilation step into the
expression function.