Revision
I made the lexer be able to tell the difference between doubles and
integers.
I made it so that the generator now generates double friendly code,
this the generator is not very efficient with code space or speed.
I also added in a little data buffer so that data (IEE encoding of
floats / doubles) is stored there and then put on the end of the code
buffer. I will also work out a little reference filler function.