Doubles are difficult
I have been trying to implement doubles for a long time now.
They are still not working, but I feel like I am closer to achieving it.
I've been trying to append the double data to the end of the
code buffer, and then using RCX as an offset to it, sort
of a makeshift data segment, but it is still not working
because it is giving me a segmentation fault.
I'm not entirely sure where it fails but I would guess on the first
movsd. I will work more on fixing this tomorrow.