It works
I basically had to make a custom run function which would set the RCX
to the start of the data section and the built machine code would
revolve around the fact that the RCX register
is its data segment.
I just now have to implement the rest of the arithmetic operations and
ensure that register allocation works.