Variables
I am currently trying to get variables to work within the
code generator. I have had to make some comprimises on the interface
of the library, by using arrays and I think I will have to add in one
more argument to the expression and run functions which is the length
of the arrays. I would like to correct this back into the original
design in the future, but for now this is easier.
I have the code generation for variables working, all I need to do is
to get the variables pushing to the stack right.