Implemented variables
I needed to change subtraction to addition, and
the first variable is in the 0x18 spot.
That's about it, I changed the interface more to be easier to program.
I believe now it is my obligation to clean up this code, it is getting
very messy and the documentation is a bit out of date.