Cleaning up
I cleaned up the comments, code, and the output. I've also reworked the
interface to now no longer require the use of arrays, rather in the
expression function the programmer specifies the number of variables
and the variable names, and then in the run function the number of
variables no longer have to be passed.