Derivative work
Taking the derivative of only one term works for now
but I have an idea. I am making this commit as a back up.
Code generation was fixed, that was because the parser was
not setting the type of the new_node, and the exponent
generator also had a problem in it where it was multiplying
by one extra. I see a problem for exponents that go above 2, but I will
fix this in a later commit.