Custom structures
I made it so that the default structures are able to specify a custom
structure, this is not the most efficient way of doing it, but it is
quite simple for the end user.
I don't think there is anyway to make it so that stdarg sees all the
arguments of a function as a certain type, so unfortunately the TODO
in test.c could not be fufilled.